requirements.txt
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.9.5
django-cors-headers==1.1.0
django-extensions==1.6.3
djangorestframework==3.3.3
oauthlib==1.1.0
psycopg2==2.6.1
requests==2.10.0
requests-oauthlib==0.6.1
six==1.10.0
static3==0.6.1
tweepy==3.5.0
